feat: coment out components except Drawer
Material Tailwind has a bug which force to me to set over 250 parameters when deploying on Amplify.
Alternatively, NextUI can be used, but it does not have Drawer component.
So commenting out components except Drawer, and try to see if deployment goes well.
If so, keep the Material Tailwind for Drawer, and replace other components with NextUI.